Task # 15 Refactored Configuration files
